Reinforcement Learning/requirements.txt
# Core dependencies
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.2
matplotlib==3.8.3

# Deep Learning
tensorflow==2.16.1

# Reinforcement Learning
gymnasium==0.29.1  # Modern replacement for gym

# Utilities
pyyaml==6.0.1
tqdm==4.66.2
python-dotenv==1.0.1
